Moved energy splitting functions into utils
